Homepage: Add hover emoji animation to the hero section
